upd input handling method to allow seamless implementation for keyboard/mouse and controller
