+++ Ravel Engine/Source/Controller/PlayerController.cpp
@@ -39,16 +39,13 @@ void PlayerController::LoadFromFile(std::string const& file) {
 
 void PlayerController::Update() {
 	// trap keys
-	if (OnKeyHold(VK_LEFT)) {
-		pawn->SetPosition(pawn->position + pawn->right * -movement_speed);
-	}
-	if (OnKeyHold(VK_RIGHT)) {
-		pawn->SetPosition(pawn->position + pawn->right * movement_speed);
-	}
 	if (OnKeyHold(VK_UP)) {
 		pawn->SetPosition(pawn->position + pawn->up * movement_speed);
 	}
 	if (OnKeyHold(VK_DOWN)) {
 		pawn->SetPosition(pawn->position + pawn->up * -movement_speed);
 	}
+
+	float delta = Input::GetAxis("Horizontal") * movement_speed;
+	pawn->SetPosition(pawn->position + pawn->right * delta);
 }
