+++ Ravel Engine/Source/Systems/Core/RavelEngine.cpp
@@ -108,6 +108,7 @@ void RavelEngine::SystemInit(HINSTANCE hInstance, int nCmdShow) {
 	SceneManagement()->Init();
 
 	AudioManagement()->Init();
+	GetInput()->InitializeInput(GetWindowHandle());
 
 #ifdef _DEBUG
 	std::cout << "Done!" << std::endl;
